docker_odoo_rfid_mqtt/requirements.txt
paho-mqtt==1.2.3
pycryptodome==3.4.11
